+++ apps/nextjs-example/src/components/WalletSelector.tsx
@@ -2,16 +2,26 @@
 
 import {
   APTOS_CONNECT_ACCOUNT_URL,
+  AboutAptosConnect,
+  AboutAptosConnectEducationScreen,
   AnyAptosWallet,
+  AptosPrivacyPolicy,
   WalletItem,
-  getAptosConnectWallets,
+  WalletSortingOptions,
+  groupAndSortWallets,
   isAptosConnectWallet,
   isInstallRequired,
-  partitionWallets,
   truncateAddress,
   useWallet,
 } from "@aptos-labs/wallet-adapter-react";
-import { ChevronDown, Copy, LogOut, User } from "lucide-react";
+import {
+  ArrowLeft,
+  ArrowRight,
+  ChevronDown,
+  Copy,
+  LogOut,
+  User,
+} from "lucide-react";
 import { useCallback, useState } from "react";
 import { Button } from "./ui/button";
 import {
@@ -34,7 +44,7 @@ import {
 } from "./ui/dropdown-menu";
 import { useToast } from "./ui/use-toast";
 
-export function WalletSelector() {
+export function WalletSelector(walletSortingOptions: WalletSortingOptions) {
   const { account, connected, disconnect, wallet } = useWallet();
   const { toast } = useToast();
   const [isDialogOpen, setIsDialogOpen] = useState(false);
@@ -91,77 +101,97 @@ export function WalletSelector() {
       <DialogTrigger asChild>
         <Button>Connect a Wallet</Button>
       </DialogTrigger>
-      <ConnectWalletDialog close={closeDialog} />
+      <ConnectWalletDialog close={closeDialog} {...walletSortingOptions} />
     </Dialog>
   );
 }
 
-interface ConnectWalletDialogProps {
+interface ConnectWalletDialogProps extends WalletSortingOptions {
   close: () => void;
 }
 
-function ConnectWalletDialog({ close }: ConnectWalletDialogProps) {
+function ConnectWalletDialog({
+  close,
+  ...walletSortingOptions
+}: ConnectWalletDialogProps) {
   const { wallets = [] } = useWallet();
 
-  const {
-    /** Wallets that use social login to create an account on the blockchain */
-    aptosConnectWallets,
-    /** Wallets that use traditional wallet extensions */
-    otherWallets,
-  } = getAptosConnectWallets(wallets);
+  const { aptosConnectWallets, availableWallets, installableWallets } =
+    groupAndSortWallets(wallets, walletSortingOptions);
 
-  const {
-    /** Wallets that are currently installed or loadable. */
-    defaultWallets,
-    /** Wallets that are NOT currently installed or loadable. */
-    moreWallets,
-  } = partitionWallets(otherWallets);
+  const hasAptosConnectWallets = !!aptosConnectWallets.length;
 
   return (
     <DialogContent className="max-h-screen overflow-auto">
-      <DialogHeader className="flex flex-col items-center">
-        <DialogTitle className="flex flex-col text-center leading-snug">
-          <span>Log in or sign up</span>
-          <span>with Social + Aptos Connect</span>
-        </DialogTitle>
-      </DialogHeader>
-      <div className="flex flex-col gap-3 pt-3">
-        {aptosConnectWallets.map((wallet) => (
-          <AptosConnectWalletRow
-            key={wallet.name}
-            wallet={wallet}
-            onConnect={close}
-          />
-        ))}
-      </div>
-      <div className="flex items-center gap-3 pt-4 text-muted-foreground">
-        <div className="h-px w-full bg-secondary" />
-        Or
-        <div className="h-px w-full bg-secondary" />
-      </div>
-      <div className="flex flex-col gap-3 pt-3">
-        {defaultWallets.map((wallet) => (
-          <WalletRow key={wallet.name} wallet={wallet} onConnect={close} />
-        ))}
-        {!!moreWallets.length && (
-          <Collapsible className="flex flex-col gap-3">
-            <CollapsibleTrigger asChild>
-              <Button size="sm" variant="ghost" className="gap-2">
-                More wallets <ChevronDown />
-              </Button>
-            </CollapsibleTrigger>
-            <CollapsibleContent className="flex flex-col gap-3">
-              {moreWallets.map((wallet) => (
-                <WalletRow
-                  key={wallet.name}
-                  wallet={wallet}
-                  onConnect={close}
-                />
-              ))}
-            </CollapsibleContent>
-          </Collapsible>
+      <AboutAptosConnect renderEducationScreen={renderEducationScreen}>
+        <DialogHeader>
+          <DialogTitle className="flex flex-col text-center leading-snug">
+            {hasAptosConnectWallets ? (
+              <>
+                <span>Log in or sign up</span>
+                <span>with Social + Aptos Connect</span>
+              </>
+            ) : (
+              "Connect Wallet"
+            )}
+          </DialogTitle>
+        </DialogHeader>
+
+        {hasAptosConnectWallets && (
+          <div className="flex flex-col gap-2 pt-3">
+            {aptosConnectWallets.map((wallet) => (
+              <AptosConnectWalletRow
+                key={wallet.name}
+                wallet={wallet}
+                onConnect={close}
+              />
+            ))}
+            <p className="flex gap-1 justify-center items-center text-muted-foreground text-sm">
+              Learn more about{" "}
+              <AboutAptosConnect.Trigger className="flex gap-1 py-3 items-center text-foreground">
+                Aptos Connect <ArrowRight size={16} />
+              </AboutAptosConnect.Trigger>
+            </p>
+            <AptosPrivacyPolicy className="flex flex-col items-center py-1">
+              <p className="text-xs leading-5">
+                <AptosPrivacyPolicy.Disclaimer />{" "}
+                <AptosPrivacyPolicy.Link className="text-muted-foreground underline underline-offset-4" />
+                <span className="text-muted-foreground">.</span>
+              </p>
+              <AptosPrivacyPolicy.PoweredBy className="flex gap-1.5 items-center text-xs leading-5 text-muted-foreground" />
+            </AptosPrivacyPolicy>
+            <div className="flex items-center gap-3 pt-4 text-muted-foreground">
+              <div className="h-px w-full bg-secondary" />
+              Or
+              <div className="h-px w-full bg-secondary" />
+            </div>
+          </div>
         )}
-      </div>
+
+        <div className="flex flex-col gap-3 pt-3">
+          {availableWallets.map((wallet) => (
+            <WalletRow key={wallet.name} wallet={wallet} onConnect={close} />
+          ))}
+          {!!installableWallets.length && (
+            <Collapsible className="flex flex-col gap-3">
+              <CollapsibleTrigger asChild>
+                <Button size="sm" variant="ghost" className="gap-2">
+                  More wallets <ChevronDown />
+                </Button>
+              </CollapsibleTrigger>
+              <CollapsibleContent className="flex flex-col gap-3">
+                {installableWallets.map((wallet) => (
+                  <WalletRow
+                    key={wallet.name}
+                    wallet={wallet}
+                    onConnect={close}
+                  />
+                ))}
+              </CollapsibleContent>
+            </Collapsible>
+          )}
+        </div>
+      </AboutAptosConnect>
     </DialogContent>
   );
 }
@@ -205,5 +235,55 @@ function AptosConnectWalletRow({ wallet, onConnect }: WalletRowProps) {
         </Button>
       </WalletItem.ConnectButton>
     </WalletItem>
+  );
+}
+
+function renderEducationScreen(screen: AboutAptosConnectEducationScreen) {
+  return (
+    <>
+      <DialogHeader className="grid grid-cols-[1fr_4fr_1fr] items-center space-y-0">
+        <Button variant="ghost" size="icon" onClick={screen.cancel}>
+          <ArrowLeft />
+        </Button>
+        <DialogTitle className="leading-snug text-base text-center">
+          About Aptos Connect
+        </DialogTitle>
+      </DialogHeader>
+
+      <div className="flex h-[162px] pb-3 items-end justify-center">
+        <screen.Graphic />
+      </div>
+      <div className="flex flex-col gap-2 text-center pb-4">
+        <screen.Title className="text-xl" />
+        <screen.Description className="text-sm text-muted-foreground [&>a]:underline [&>a]:underline-offset-4 [&>a]:text-foreground" />
+      </div>
+
+      <div className="grid grid-cols-3 items-center">
+        <Button
+          size="sm"
+          variant="ghost"
+          onClick={screen.back}
+          className="justify-self-start"
+        >
+          Back
+        </Button>
+        <div className="flex items-center gap-2 place-self-center">
+          {screen.screenIndicators.map((ScreenIndicator, i) => (
+            <ScreenIndicator key={i} className="py-4">
+              <div className="h-0.5 w-6 transition-colors bg-muted [[data-active]>&]:bg-foreground" />
+            </ScreenIndicator>
+          ))}
+        </div>
+        <Button
+          size="sm"
+          variant="ghost"
+          onClick={screen.next}
+          className="gap-2 justify-self-end"
+        >
+          {screen.screenIndex === screen.totalScreens - 1 ? "Finish" : "Next"}
+          <ArrowRight size={16} />
+        </Button>
+      </div>
+    </>
   );
 }