+++ packages/wallet-adapter-mui-design/src/WalletConnector.tsx
@@ -1,28 +1,21 @@
-import { AnyAptosWallet } from "@aptos-labs/wallet-adapter-react";
+import { WalletSortingOptions } from "@aptos-labs/wallet-adapter-react";
+import { Breakpoint } from "@mui/material";
 import { useState } from "react";
 import WalletButton from "./WalletButton";
 import WalletsModal from "./WalletModel";
 
-export interface WalletConnectorProps {
+export interface WalletConnectorProps extends WalletSortingOptions {
   networkSupport?: string;
   handleNavigate?: () => void;
-  /**
-   * An optional function for sorting wallets that are currently installed or
-   * loadable in the wallet connector modal.
-   */
-  sortDefaultWallets?: (a: AnyAptosWallet, b: AnyAptosWallet) => number;
-  /**
-   * An optional function for sorting wallets that are NOT currently installed or
-   * loadable in the wallet connector modal.
-   */
-  sortMoreWallets?: (a: AnyAptosWallet, b: AnyAptosWallet) => number;
+  /** The max width of the wallet selector modal. Defaults to `xs`. */
+  modalMaxWidth?: Breakpoint;
 }
 
 export function WalletConnector({
   networkSupport,
   handleNavigate,
-  sortDefaultWallets,
-  sortMoreWallets,
+  modalMaxWidth,
+  ...walletSortingOptions
 }: WalletConnectorProps) {
   const [modalOpen, setModalOpen] = useState(false);
   const handleModalOpen = () => setModalOpen(true);
@@ -38,8 +31,8 @@ export function WalletConnector({
         handleClose={handleClose}
         modalOpen={modalOpen}
         networkSupport={networkSupport}
-        sortDefaultWallets={sortDefaultWallets}
-        sortMoreWallets={sortMoreWallets}
+        modalMaxWidth={modalMaxWidth}
+        {...walletSortingOptions}
       />
     </>
   );