+++ packages/wallet-adapter-core/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @aptos-labs/wallet-adapter-core
 
+## 4.9.0
+
+### Minor Changes
+
+- 96df1f7: Added `groupAndSortWallets` utility function.
+- f23cf43: Fix adapter event communication
+
+## 4.8.2
+
+### Patch Changes
+
+- 79a0212: Bumped to `@aptos-connect/wallet-adapter-plugin` v1.0.1
+
+## 4.8.1
+
+### Patch Changes
+
+- 1644cfc: use isAptosConnectWallet function to verify a wallet is an AptosConnect wallet in excludeWallet function
+- cbbbe23: Added Dapp id to dappConfig
+
+## 4.8.0
+
+### Minor Changes
+
+- e3df2db: Added the following utility functions: `getAptosConnectWallets`, `partitionWallets`, `isInstalledOrLoadable`, `isInstallRequired`, and `truncateAddress`.
+- 1580df8: Add Petra to the AIP-62 wallet registry
+
 ## 4.7.0
 
 ### Minor Changes